Allow mint submissions as start block is pending
Current implementation only allows mint submissions after the start block is confirmed meaning ord users can not attempt to compete in that block. This implementation allows for submissions at start - 1, which is a good solution for most users. However, a flag to ignore these checks for power users might be a more composable solution.