Update specification.md
What happens if we try to allocate 10 runes to each of 3 non op_return outputs (by setting `output` to the number of outputs) when there are only 17 runes available to allocate? 

What happens is that: 

* The first output gets 10,
* the 2nd gets 7, and
* the 3rd gets 0.

This isn't explicitly covered in the spec. Adding "in order" hints at how this works. Is that enough? Or does the spec need to be explicit about this?