"Add gt_sequence for export function and refactor code"
In this commit, we have added the "gt_sequence" argument for the "export" function to enable the user to export from a specific sequence number. Also, redundant code pieces and the "rune" related functions have been removed. This refactor brings overall simplification to the codebase. Modifications have been made in various areas, including Inscription structures and dependencies in the Cargo.toml file.